move token functionlities to EmailVerificationIndex
